requirements.txt
celery
redis
aiofiles
alembic
asyncio
asyncpg
elasticsearch
elasticsearch_dsl
fastapi
passlib
psycopg2-binary
pydantic
PyJWT
python-dotenv
SQLAlchemy
starlette
typing_extensions
uvicorn
python-multipart
aiohttp
fastapi-mail
requests